+++ src/DbTests.DbGen/IncludeFixtures/Count.sql
@@ -1,0 +1,2 @@
+DECLARE @countOnly BIT = 1;
+{{include "Shared/Search.sql"}}